requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
coverage==4.5.2
docopt==0.6.2
Flask==1.0.2
Flask-Script==2.0.6
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
more-itertools==5.0.0
paho-mqtt==1.3.1
pbr==5.1.1
pluggy==0.8.1
py==1.7.0
PyJWT==1.7.1
pykwalify==1.7.0
pytest==4.1.1
pytest-cov==2.6.1
pytest-flask==0.14.0
python-box==3.2.3
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
six==1.12.0
stevedore==1.30.0
tavern==0.22.1
urllib3==1.24.1
Werkzeug==0.14.1
vcrpy==1.7.4
